Fix numeric args to create, login, monitor-events commands
The create, login, and monitor-events commands were interpreting certain numeric args as strings instead of integers, which, when specified, caused the commands to fail (actually, create happened to work but I updated it for consistency). With this fix the arguments are now parsed as integers, so you can: specify a --daysValid value when creating a device with an X509 cert; specify a --duration value when using "login" to start a session; specify a --start-time (as milliseconds since epoch OR as ISO-8601 string--I verified both) when monitoring events for a device.

The --duration arg to sas-token was already correct, but I updated it for consistency. I believe the "best practice" is to pass the parsing function to commander.option() if possible, then fall back to manual parsing if needed (monitor-events --start-time is an example of an overloaded arg that needs to be parsed manually).